+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 **Changed:**
 - DNS Service - DNS Service moved from API calls to Azure latest SDK.
 
+
+
 ## 0.0.8 
 
 **Added:**
@@ -22,6 +24,7 @@
     get blob properties, set blob metadata, get blob metadata, get blob, list blob functionality for storage
 - Added get container properties, get container metadata, set container metadata, get container, list container,
     delete container,  Get the access control list of the container
+
 
 ## 0.0.6